requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.3
line-bot-sdk>=2.4.2